+++ core/bridge.js
@@ -16,7 +16,7 @@ const config = require('./shared/config');
 const logging = require('@tryghost/logging');
 const tpl = require('@tryghost/tpl');
 const themeEngine = require('./frontend/services/theme-engine');
-const frontendRouting = require('./frontend/services/routing').bootstrap;
+const routerManager = require('./frontend/services/routing').routerManager;
 const settingsCache = require('./shared/settings-cache');
 
 // Listen to settings.lang.edited, similar to the member service and models/base/listeners
@@ -41,7 +41,7 @@ class Bridge {
         //       for now this eliminates the need for the frontend routing to listen to
         //       server events
         events.on('settings.timezone.edited', (model) => {
-            frontendRouting.handleTimezoneEdit(model);
+            routerManager.handleTimezoneEdit(model);
         });
     }
