+++ core/frontend/services/routing/controllers/email-post.js
@@ -1,6 +1,6 @@
 const debug = require('@tryghost/debug')('services:routing:controllers:emailpost');
 const config = require('../../../../shared/config');
-const {bootstrap} = require('../');
+const {routerManager} = require('../');
 const urlUtils = require('../../../../shared/url-utils');
 const helpers = require('../helpers');
 
@@ -48,7 +48,7 @@ module.exports = function emailPostController(req, res, next) {
             }
 
             if (post.status === 'published') {
-                return urlUtils.redirect301(res, bootstrap.getUrlByResourceId(post.id, {withSubdirectory: true}));
+                return urlUtils.redirect301(res, routerManager.getUrlByResourceId(post.id, {withSubdirectory: true}));
             }
 
             if (res.locals.apiVersion !== 'v0.1' && res.locals.apiVersion !== 'v2') {
