+++ test/unit/frontend/services/routing/controllers/collection.test.js
@@ -4,7 +4,7 @@ const sinon = require('sinon');
 const testUtils = require('../../../../../utils');
 const security = require('@tryghost/security');
 const themeEngine = require('../../../../../../core/frontend/services/theme-engine');
-const bootstrap = require('../../../../../../core/frontend/services/routing/').bootstrap;
+const routerManager = require('../../../../../../core/frontend/services/routing/').routerManager;
 const controllers = require('../../../../../../core/frontend/services/routing/controllers');
 const helpers = require('../../../../../../core/frontend/services/routing/helpers');
 
@@ -56,7 +56,7 @@ describe('Unit - services/routing/controllers/collection', function () {
 
         sinon.stub(helpers, 'renderEntries').returns(renderStub);
 
-        ownsStub = sinon.stub(bootstrap, 'owns');
+        ownsStub = sinon.stub(routerManager, 'owns');
         ownsStub.withArgs('identifier', posts[0].id).returns(true);
 
         req = {
