+++ test/unit/frontend/services/rss/generate-feed.test.js
@@ -3,7 +3,7 @@ const sinon = require('sinon');
 const _ = require('lodash');
 const testUtils = require('../../../../utils');
 const configUtils = require('../../../../utils/configUtils');
-const bootstrap = require('../../../../../core/frontend/services/routing').bootstrap;
+const routerManager = require('../../../../../core/frontend/services/routing').routerManager;
 const generateFeed = require('../../../../../core/frontend/services/rss/generate-feed');
 
 describe('RSS: Generate Feed', function () {
@@ -43,7 +43,7 @@ describe('RSS: Generate Feed', function () {
     });
 
     beforeEach(function () {
-        sinon.stub(bootstrap, 'getUrlByResourceId');
+        sinon.stub(routerManager, 'getUrlByResourceId');
 
         baseUrl = '/rss/';
 
@@ -92,7 +92,7 @@ describe('RSS: Generate Feed', function () {
             data.posts = posts;
 
             _.each(data.posts, function (post) {
-                bootstrap.getUrlByResourceId.withArgs(post.id, {secure: undefined, absolute: true}).returns('http://my-ghost-blog.com/' + post.slug + '/');
+                routerManager.getUrlByResourceId.withArgs(post.id, {secure: undefined, absolute: true}).returns('http://my-ghost-blog.com/' + post.slug + '/');
             });
 
             generateFeed(baseUrl, data).then(function (xmlData) {
@@ -204,7 +204,7 @@ describe('RSS: Generate Feed', function () {
             data.posts = [posts[0]];
 
             _.each(data.posts, function (post) {
-                bootstrap.getUrlByResourceId.withArgs(post.id, {secure: undefined, absolute: true}).returns('http://my-ghost-blog.com/' + post.slug + '/');
+                routerManager.getUrlByResourceId.withArgs(post.id, {secure: undefined, absolute: true}).returns('http://my-ghost-blog.com/' + post.slug + '/');
             });
 
             generateFeed(baseUrl, data).then(function (xmlData) {
